cmms/requirements.txt
# To activate this run
#
# pip install -r requirements.txt
# python -m django --version
#
# On a Macintosh this should be python3 and pip3
#
Django>=3.1.4
django-crispy-forms>=1.8.1
django-filter>=2.2.0
djangorestframework>=3.10.3
Markdown>=3.0.1
social-auth-app-django>=3.1.0
social-auth-core>=3.3.3
django-extensions>=2.2.5
django-taggit>=1.3.0
mysqlclient>=1.4.6
psycopg2
django-rosetta
django-modeltranslation>=0.17.3
django-right-to-left>=0.1.1

